+++ src/functions.js
@@ -1,6 +1,4 @@
 import { promises as fsPromises } from 'fs';
-import { DiacriticsMap } from './constants.js';
-
 class Functions {
     constructor() {
     }
@@ -43,28 +41,6 @@ class Functions {
             wrappedError.original = error;
             throw wrappedError;
         }
-    }
-
-    async sanitizeString(str) {
-        if (!str) return '';
-
-        // Replace diacritics using map
-        str = str.replace(/[^\u0000-\u007E]/g, ch => DiacriticsMap[ch] || ch);
-
-        // Replace separators between words with space
-        str = str.replace(/(\w)[.:;+\-\/]+(\w)/g, '$1 $2');
-
-        // Replace remaining standalone separators with space
-        str = str.replace(/[.:;+\-\/]/g, ' ');
-
-        // Remove remaining invalid characters (keep letters, digits, space, apostrophe)
-        str = str.replace(/[^A-Za-z0-9 ']/g, ' ');
-
-        // Collapse multiple spaces
-        str = str.replace(/\s+/g, ' ');
-
-        // Trim
-        return str.trim();
     }
 
     async findIfaceBySsid(status, targetSsid) {